Better filtering of noise from JVM standard error stream.
The environment variable _JAVA_OPTIONS, and the similar
JAVA_TOOL_OPTIONS, can be used to set global options for the Java
virtual machine.  Unfortunately, if these variables are set, every
invocation of the JVM will then display a message to standard error
which cannot be suppressed.

If we are running the validator in JSON (or XML) mode, these messages
must be removed before attempting to parse the output.  Thus, it makes
sense to filter these out at the level of _run_validator rather than
by treating them as "error messages" to be ignored.

In addition, make the filtering a bit more robust: only match these
messages at the beginning of a line; and handle JAVA_TOOL_OPTIONS as
well as _JAVA_OPTIONS.